perf(db): big payloads last in row + feed covering index; finalize cutover
SQLite inlines large values in the row record, and reading any column
stored AFTER them walks their overflow pages. items carried body (~KBs),
body_md (~30KB) and embedding (12KB) mid-row, so every feed filter on
cluster_id/tier paid a ~45KB-per-row overflow walk: the joined feed count
measured 29.4s. Fix:

- db/schema.ts: body/body_md/embedding moved to the END of items (comment
  documents the invariant); raw_payload moved last in raw_items.
- items_feed_cover_idx (enriched_at, importance, tier, cluster_id,
  source_id, published_at) — feed/count scans use the index, not the fat
  rows. Measured after rebuild: count 29.4s -> 182ms, feed page 3.2s ->
  0.6s, usage summary 38ms.
- Live v2 DB rebuilt in place (chunked server-side INSERT INTO ... SELECT,
  swap, indexes recreated; counts + FK check verified).
- Tests: remaining pg-era assertions updated (pgEnum -> text enums,
  ::int casts, ms fixtures); public-search pagination test gets a 20s
  timeout (two sequential remote LIKE scans, RTT-bound from dev machines).
- One-off migration/verify scripts removed; `postgres` dependency removed.
- bun run verify fully green: typecheck, lint, build, knip x3, 1038 tests.

+++ tests/api/saved-routes.test.ts
@@ -64,7 +64,7 @@ describeOrSkip("saveItemRoutePayload (real DB)", () => {
 
     const [maxItem] = await db()
       .select({
-        maxId: sql<number>`coalesce(max(${schema.items.id}), 0)::int`,
+        maxId: sql<number>`coalesce(max(${schema.items.id}), 0)`,
       })
       .from(schema.items);
     missingItemId = (maxItem?.maxId ?? 0) + 1_000_000;
